// src/types/battery-transaction.types.ts
// Battery Transaction Log Information
export interface BatteryTransactionLog {
  createBy: string;
  createTime: string;
  updateBy: string;
  updateTime: string;
  logId: string;
  logType: string;
  logStatus: string;
  logStatusMsg?: string;
  orderId: string;
  userId: number;
  switchCabId: string;
  reqTime: string;
  resTime: string;
  successTime?: string;
  failTime?: string;
  overTime?: string;
  getBatteryId: string;
  getBatterySoc: number;
  getBatteryRemainCapacity: number;
  getBatteryDoorId: number;
  putBatteryId: string;
  putBatterySoc: number;
  putBatteryRemainCapacity: number;
  putBatteryDoorId: number;
  flowId: string;
  flowType: string;
  balanceAmount: number;
  virtualAmount: number;
  depositAmount: number;
  balanceAccount: number;
  virtualAccount: number;
  depositAccount: number;
  flowOrderId: string;
  flowCreateBy: string;
  flowCreateTime: string;
  flowUpdateBy?: string;
  userName: string;
  phoneNumber: string;
  deptName: string;
  mark: number;
  batteryCount?: number;
  deptId?: number;
  num?: number;
  rang?: string;
  remark?: string;
  time?: string;
  shardingKey?: number;
  params?: Record<string, any>;
}

// Query Parameters based on Swagger
export interface BatteryTransactionQuery {
  batteryCount?: number;
  createBy?: string;
  createTime?: string;
  deptId?: number;
  failTime?: string;
  getBatteryDoorId?: number;
  getBatteryId?: string;
  getBatteryRemainCapacity?: number;
  getBatterySoc?: number;
  logId?: number;
  logStatus?: string;
  logStatusMsg?: string;
  logType?: string;
  num?: number;
  orderId?: number;
  overTime?: string;
  params?: Record<string, any>;
  putBatteryDoorId?: number;
  putBatteryId?: string;
  putBatteryRemainCapacity?: number;
  putBatterySoc?: number;
  rang?: string;
  remark?: string;
  reqTime?: string;
  resTime?: string;
  searchValue?: string;
  shardingKey?: number;
  successTime?: string;
  switchCabId?: string;
  time?: string;
  updateBy?: string;
  updateTime?: string;
  userId?: number;
  page?: number;
  size?: number;
  searchText?: string;
}